TOTS-57 Removed some code duplication around the "onClick" parameter to the "MapPopup" component.
